#198 [리뷰반영] 패키지 구조 수정 및 파일분리
- 'listitem' 패키지를 제거하고 존재하던 파일들을 상위 패키지로 이동합니다
- BookmarkItemUiState를 파일로 분리합니다
